Software-GUI/requirements.txt
pygame-ce>=2.5.7
python-chess>=1.999
